+++ bin/gamedig.js
@@ -6,8 +6,11 @@ import Minimist from 'minimist'
 import { GameDig } from './../lib/index.js'
 
 const argv = Minimist(process.argv.slice(2), {
-  boolean: ['pretty', 'debug', 'givenPortOnly', 'requestRules', 'requestRulesRequired', 'requestPlayersRequired'],
-  string: ['guildId', 'listenUdpPort', 'ipFamily']
+  boolean: ['pretty', 'debug', 'givenPortOnly', 'requestRules', 'requestRulesRequired', 'requestPlayersRequired', 'stripColors'],
+  string: ['guildId', 'listenUdpPort', 'ipFamily'],
+  default: {
+    stripColors: true
+  }
 })
 
 const options = {}